Fixing image gallery layout, errors from validator, and organizing stylesheet
Setting layout constraits for the layouts of each subject gallery. Reoranged the stylesheet more for readability. Added opacity to banner image. Fixed errors/warnings from validator on the gallery and caption section tags and changed them to div tags, because it complained about heading tags not being in the sections.